in get_rts.py, removed peak_separation parameter, I do not plan to use it. Also added a function to get from data to pixels to be passed through the get_rts function.
Made new file called test_data.py. The code to generate a 25x25 grid of pixels with fake RTS pixels is there, not functionalized. Probably want to change this to a csv file to be read in at some point. The code to test get_rts is temporarily stored here.